+++ frontend/altinn-support-dashboard.client/src/components/Dashboard/MainContentComponent.test.tsx
