+++ frontend/altinn-support-dashboard.client/src/components/Dashboard/hooks/useAuthorizedFetch.ts
