+++ frontend/altinn-support-dashboard.client/src/components/Dashboard/models/mainContentTypes.ts
