+++ frontend/altinn-support-dashboard.client/src/components/Dashboard/utils/dateUtils.ts
